srv: handle missing air quality data
It's possible that no valid air quality data exists, for example because
all data sources are stale or no data sources are defined. In that case,
there is no AQI data to aggregate (by picking the max value). Set the
corresponding memcache value to null (which is a valid JSON value), and
do not send an update (for the AQI value) to carbon cache.